Fix access of randomuser.me api property data
The modal that opens with details about a single user was broken
because the 'street' property was not properly been accessed/read.
It seems that the problem was that it used to be a string and they
decided to spited into an object, with 'number' and 'name' properties.
This commit change the code to access the properties individually.